feat: Add passthrough mode to API (#72)
* feat: Add passthrough mode to API (#passthrough)

Added a passthrough mode to API where the request body is returned unchanged if the 'X-Passthrough' header is set. Updated the README.md with example of using the new feature. Also, updated the verify-api.yaml to include a check for the passthrough mode.

* feat(tests): update test expectations to use toMatchObject #enhanced-preprints-biorxiv-xslt

Changed assertion methods in app.test.ts and transform.test.ts from toStrictEqual to toMatchObject to allow for more flexible test cases.