feat(Dockerfile): add health check to Dockerfile (#73)
* feat(Dockerfile): add health check to Dockerfile

This commit adds a health check to the Dockerfile. The health check is set to run every 30 seconds with a timeout of 30 seconds. If the health check fails 3 times, it will mark the container as unhealthy.

* feat: add health check for API container

Added a step in the GitHub Actions workflow to verify that the API
container is healthy before proceeding with further steps. This is
done by inspecting the container's health status. If it's not healthy,
the workflow will be terminated with an error exit code.

* feat: update docker run command in verify-api workflow

Removed sleep command and added container name to docker run command in .github/workflows/verify-api.yaml. This helps to make the setup more robust and easier to manage.

* feat: Improve API health check in verify-api workflow

This commit enhances the way we verify API health in the github actions
workflow. Now, instead of just waiting for a fixed time and checking
the API health, we repeatedly check the API health over a given time
interval until it becomes healthy. This should improve reliability
of the API health check.

* feat: Improve API health check messages in verify-api.yaml workflow

In the GitHub Actions workflow for verifying the API, the messages
displayed when waiting for the API to become healthy or when the API
is not healthy after a certain number of attempts have been improved
for better clarity and accuracy.